agent-engine/requirements.txt
# Core AI and Agent Framework
crewai==0.35.8
# The LangChain family has been aligned to the 0.1.x series for crewai compatibility
langchain==0.1.20
langchain-google-genai==1.0.4
langchain-community==0.0.38
langchain-core==0.1.52

# API and Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
httpx==0.27.0

# Database and Vector Storage
# Unpinned chromadb to resolve dependency conflict with crewai -> embedchain
chromadb

# Database and ORM
sqlalchemy==2.0.30
alembic==1.13.1

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# File and System Operations
watchdog==4.0.0

# Development and Testing
pytest==8.2.0
pytest-asyncio==0.23.6
black==24.4.2
isort==5.13.2
flake8==7.0.0

# Logging and Monitoring
loguru==0.7.2
rich==13.7.1

# Additional Utilities
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.9